AI_FACE/requirements.txt
fastapi
uvicorn
python-dotenv
numpy
opencv-python
imutils
face_recognition
python-multipart
SQLAlchemy
pymysql
cryptography